==> 0.packages.R <==
# Load our packages in the usual way.
# This example requires packages forcats, readxl, and rmarkdown,
# but you do not need to load them here.

library(distances)
library(drake)
library(dplyr)
library(tidyverse)
library(broom)
library(sp)
library(gstat)
library(rgdal)
library(raster)
#devtools::install_github("tidyverse/tidyr") 
library(tidyr)
library(ggplot2)
library(spBayes)
library(metR)
library(fda)
library(fda.usc)
library(tidyverse)
library(lubridate)
library(gganimate)
library(magick)
library(ggridges)
library(tidyverse)
library(hrbrthemes)
library(viridis)
library(gridExtra)
library(cowplot)
library(boot)
library(tmap)
library(scales)
library(lattice)
library(ggalt)
library(boot)
library(Kendall)
